requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.2
alembic==1.13.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==25.3.0
Babel==2.9.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.110.2
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
idna==3.10
Jinja2==3.1.3
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.5.0
propcache==0.3.2
psycopg2-binary==2.9.9
pydantic==2.11.7
pydantic_core==2.33.2
pypng==0.20220715.0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
qrcode==7.4.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.37.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.29.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
